Destroy unsupported blocks instead of changing them.
Put the block updates in batches, once per tick, so as to avoid the
removal of blocks by the FallingBlockSystem itself causing more
updates. This also allows for more consistent behaviour if other things
cause many block changes at once.

Add tests for consistency of the octree state. It's still buggy, but
this should help isolate the causes.