Replace null components with actual components. General refactoring.
Make the datastructure more consistent by making the components in
size-1 SolidNodes actual FullComponents, so that there are fewer
special cases and the code can be simplified. This also increases
the memory usage, but for normal shaped worlds (i.e. not having
lots of interleaved solid and non-solid blocks), the FallingBlockSystem's
memory usage should still be much less than the amount for block data.

Also rearrange a few other things to make them tidier and more readable,
and remove a bit of debug output that was accidentally left in.