Put subchain and touching chain data into big arrays
Reduce the amount of individual objects to make garbage-collection
easier by removing the sets of pairs from Chain and putting all that
data in manually managed arrays stored in the Tree instead. This
makes the risk of memory leaks much higher, as Chains must be
explicitly de-allocated. It's currently a work in progress, and
tends to allocate far too many arrays and run out of space.