Fix data consistency errors
Fix a bug where a subchain wouldn't be added because its parent field
was already set so it looked like it was already there. Fix a bug
(that probably pre-dates the addition of the IntPairSetHeap) where
FullNode.equivalentInternalNode would incorrectly record some subchains
as touching things they aren't actually touching. Add a bunch of debugging
stuff.